+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v0.0.5
+
+- Added CSS variables and link styles.
+- Added links from home to `/set`.
+- Added `getAllTournamentSets()` functionality to `/src/startgg.ts`.
+- Added a display for all sets of a tournament in `/routes/set/index.tsx`-
+- Updated styles in `/static/styles.css` and `/tailwind.config.ts`.
+
 ## v0.0.4
 
 - Added the route `/set/index.tsx`.